requirements.txt
pydantic-yaml==0.8.0
tqdm==4.64.1
pillow==9.2.0
opencv-python==4.6.0.66
torch==1.12.1
torchvision==0.13.1
pytorch-lightning==1.7.6
scipy==1.9.1
clearml==1.7.0